requirements.txt
Django>=3.2.4
djangorestframework>=3.12.4
rollbar>=0.15.1
sentry-sdk>=1.4.3
python-dotenv>=0.19.1
pydantic==1.8.2
marshmallow==3.14.1